# kubernetes/swagger-config copy.yaml
openapi: 3.0.3
info:
  title: Messenger API
  version: 1.0.1
  description: API мессенджера
servers:
  - url: /api
    description: REST API сервера
  - url: ws://localhost:8083
    description: WebSocket сервер для звонков
security:
  - BearerAuth: []
...
paths:
...
  /v1/notifications:
    get:
      tags:
        - Уведомления
      summary: Получить список уведомлений
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Номер страницы
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Количество уведомлений на странице
      responses:
        '200':
          description: Список уведомлений
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /v1/notifications/{notificationId}/read:
    post:
      tags:
        - Уведомления
      summary: Отметить уведомление как прочитанное
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
          description: ID уведомления
      responses:
        '200':
          description: Уведомление отмечено как прочитанное
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          description: ID уведомления
        type:
          type: string
          enum: [message, call, channel, friend_request]
        content:
          type: string
        data:
          type: object
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
...